chore: Update import paths for app_search_widget and custom_appbar_widget

+++ lib/view/utils/buy_product_widgets.dart
@@ -15,7 +15,7 @@ class BuyProductWidgets extends StatelessWidget {
 
   final Function() onPressed;
   final int totalProducts;
-  final int totalPrice;
+  final double totalPrice;
   final String titleButton;
 
   @override
